Patch on ilisp-devel titled Ilisp for emacs-22
This patch was posted to the ilisp-devel mailing list
https://sourceforge.net/mailarchive/forum.php?thread_name=906689B4-F4E0-48EE-A33B-B5C434944FB2%40cs.nyu.edu&forum_name=ilisp-devel
as an attachment to the following message.

From: Martin Cracauer <cracauer@cons.org>
Date: 2008-12-16 16:31
To: ilisp-devel

Ilisp folks,

Appended is a first diff that make SBCL in Ilisp snapshot work, after
a fashion.  It starts and evaluates fine, but the debugger is broken.
The debugger displays it's prompt twice, the cursor is left after the
first prompt.  Any attempt to enter things with the keyboard ends up
with a hung Lisp and the infamous
"Abort commands before interrupting top level? (y or n)" when hitting
Control-C.  The session seems nonrecoverable.

The below is an example session, the last line is being displayed
after hitting Control-C (not shown):


Starting ,sbcl ...
:cl-user - Yes, Master? (+ 1 2)

3
:cl-user - Yes, Master? (+ 1 'foo)

debugger invoked on a SIMPLE-TYPE-ERROR in thread #<THREAD "initial thread" {10\
00018D21}>:
  Argument Y is not a NUMBER: FOO

Type HELP for debugger help, or (SB-EXT:QUIT) to exit from SBCL.

restarts (invokable by number or by possibly-abbreviated name):
  0: [ABORT] Exit debugger, returning to top level.

(SB-KERNEL:TWO-ARG-+ 1 FOO)
0] a
0] a
a
1

----:**-F1  *sbcl*         All L20    (ILISP :load)-----------------------------
Abort commands before interrupting top level? (y or n)